added: download routes for files

+++ directory_reader.h
@@ -1,6 +1,8 @@
 #ifndef DIRECTORY_READER_H
 #define DIRECTORY_READER_H
 
-void generate_directory_html(const char *directory, char *html, int max_size);
+#include <stddef.h>
+
+void generate_directory_html(const char *directory, char *html, size_t max_size);
 
 #endif
